pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-reviewer"
version = "0.1.0"
description = "Claude-powered GitLab MR review agent"
requires-python = ">=3.11"
dependencies = [
    "python-gitlab>=4.6",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "mcp>=1.0",
    "click>=8.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.scripts]
claude-reviewer = "claude_reviewer.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_reviewer"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py311"
